command added

+++ app/Console/Commands/DBInsert.php
@@ -3,28 +3,28 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
-use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Artisan;
 use Madnest\Madzipper\Facades\Madzipper;
 
-class DatabaseRefresh extends Command
+class DBInsert extends Command
 {
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'database:refresh';
+    protected $signature = 'database:insert';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Refresh database after a certain time';
+    protected $description = 'Insert database after a certain time';
 
     /**
      * Create a new command instance.
@@ -43,8 +43,10 @@ class DatabaseRefresh extends Command
      */
     public function handle()
     {
-        Artisan::call('db:wipe');
-        $sql_path = base_path('installation/backup/database.sql');
-        DB::unprepared(file_get_contents($sql_path));
+        if (!Schema::hasTable('admins')) {
+            Artisan::call('db:wipe');
+            $sql_path = base_path('database/backup/database.sql');
+            DB::unprepared(file_get_contents($sql_path));
+        }
     }
 }
